refactor and comments from Mrma
