+++ tests/integration/test_memgraph.py
@@ -11,15 +11,11 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
+
 from typing import List
 
 import pytest
 from gqlalchemy import Memgraph, Node, Relationship
-
-import unittest.mock as mock
-from gqlalchemy.memgraph import QueryModule
-
-from gqlalchemy.query_builder import QueryBuilder
 
 
 def compare_nodes(actual: List[Node], expected: List[Node]):
@@ -144,78 +140,3 @@ def test_path_mapping(populated_memgraph: Memgraph):
 
     compare_nodes(actual_path._nodes, expected_nodes)
     compare_edges(actual_path._relationships, expected_relationships)
-
-
-@pytest.mark.parametrize(
-    "signature, arguments, returns",
-    [
-        ("dummy_module.1(num :: NUMBER) :: ()", [{"name": "num", "type": "NUMBER"}], []),
-        (
-            "dummy_module.2(lst :: LIST OF STRING, num = 3 :: NUMBER) :: (ret :: STRING)",
-            [{"name": "lst", "type": "LIST OF STRING"}, {"name": "num", "type": "NUMBER", "default": 3}],
-            [{"name": "ret", "type": "STRING"}],
-        ),
-    ],
-)
-def test_parse_signature(memgraph: Memgraph, signature: str, arguments: List, returns: List):
-    """test functionality of parsing a module signature"""
-    assert arguments, returns == memgraph._parse_signature(signature)
-
-
-def test_get_procedures(memgraph: Memgraph):
-    """test get procedures with mock execute_and_fetch method, so MAGE query
-    modules are not needed for testing"""
-
-    mock_modules = [
-        {
-            "is_editable": True,
-            "is_write": False,
-            "name": "max_flow.get_flow",
-            "path": "/home/bruno/mage/python/max_flow.py",
-            "signature": 'max_flow.get_flow(start_v :: NODE, end_v :: NODE, edge_property = "weight" :: STRING) :: (max_flow :: NUMBER)',
-        },
-        {
-            "is_editable": True,
-            "is_write": False,
-            "name": "max_flow.get_paths",
-            "path": "/home/bruno/mage/python/max_flow.py",
-            "signature": 'max_flow.get_paths(start_v :: NODE, end_v :: NODE, edge_property = "weight" :: STRING) :: (flow :: NUMBER, path :: PATH)',
-        },
-    ]
-
-    mock_execute_and_fetch = mock.Mock()
-    mock_execute_and_fetch.return_value = mock_modules
-
-    real_execute_and_fetch = memgraph.execute_and_fetch
-
-    def mock_execute_and_fetch_wrapper(query):
-        if query == "CALL mg.procedures() YIELD *;":
-            return mock_execute_and_fetch(query)
-        else:
-            return real_execute_and_fetch(query)
-
-    memgraph.execute_and_fetch = mock_execute_and_fetch_wrapper
-
-    assert str(memgraph.get_procedures()[0]) == "max_flow.get_flow"
-
-
-def test_query_module_with_query_builder(memgraph):
-    mock_module = {
-        "is_editable": True,
-        "is_write": False,
-        "name": "max_flow.get_flow",
-        "path": "/home/bruno/mage/python/max_flow.py",
-        "signature": 'max_flow.get_flow(start_v :: NODE, end_v :: NODE, edge_property = "weight" :: STRING) :: (max_flow :: NUMBER)',
-    }
-
-    query_module = QueryModule(mock_module)
-
-    query_module.set_inputs(start_v=None, end_v=None)
-
-    query_builder = QueryBuilder().call(procedure=query_module, arguments=query_module.get_inputs())
-    expected_query = ' CALL max_flow.get_flow(None, None, "weight") '
-
-    with mock.patch.object(Memgraph, "execute", return_value=None) as m:
-        query_builder.execute()
-
-    m.assert_called_with(expected_query)
